+++ README.md
@@ -1,8 +1,8 @@
-# HTML::Proofer
+# HTMLProofer
 
 If you generate HTML files, _then this tool might be for you_.
 
-`HTML::Proofer` is a set of tests to validate your HTML output. These tests check if your image references are legitimate, if they have alt tags, if your internal links are working, and so on. It's intended to be an all-in-one checker for your output.
+`HTMLProofer` is a set of tests to validate your HTML output. These tests check if your image references are legitimate, if they have alt tags, if your internal links are working, and so on. It's intended to be an all-in-one checker for your output.
 
 [![Build Status](https://travis-ci.org/gjtorikian/html-proofer.svg?branch=master)](https://travis-ci.org/gjtorikian/html-proofer) [![Gem Version](https://badge.fury.io/rb/html-proofer.svg)](http://badge.fury.io/rb/html-proofer)
 
@@ -62,7 +62,7 @@ You can enable or disable most of the following checks.
 
 ### Using in a script
 
-Require the gem; generate some HTML; create a new instance of the `HTML::Proofer` on
+Require the gem; generate some HTML; create a new instance of the `HTMLProofer` on
 your output folder; then `run` it. Here's a simple example:
 
 ```ruby
@@ -89,7 +89,7 @@ Find.find("./docs") do |path|
 end
 
 # test your out dir!
-HTML::Proofer.new("./out").run
+HTMLProofer.new("./out").run
 ```
 
 ### Using on the command-line
@@ -116,7 +116,7 @@ require 'html/proofer'
 
 task :test do
   sh "bundle exec jekyll build"
-  HTML::Proofer.new("./_site").run
+  HTMLProofer.new("./_site").run
 end
 ```
 
@@ -131,7 +131,7 @@ htmlproof ./_site
 Instead of a directory as the first argument, you can also pass in an array of links:
 
 ``` ruby
-HTML::Proofer.new(["http://github.com", "http://jekyllrb.com"])
+HTMLProofer.new(["http://github.com", "http://jekyllrb.com"])
 ```
 
 This configures Proofer to just test those links to ensure they are valid. Note that for the command-line, you'll need to pass a special `--as-links` argument:
@@ -150,7 +150,7 @@ Add the `data-proofer-ignore` attribute to any tag to ignore it from every check
 
 ## Configuration
 
-The `HTML::Proofer` constructor takes an optional hash of additional options:
+The `HTMLProofer` constructor takes an optional hash of additional options:
 
 | Option | Description | Default |
 | :----- | :---------- | :------ |
@@ -204,10 +204,10 @@ opts = { :check_html => true, :validation => { :ignore_script_embeds => true } }
 [Typhoeus](https://github.com/typhoeus/typhoeus) is used to make fast, parallel requests to external URLs. You can pass in any of Typhoeus' options for the external link checks with the options namespace of `:typhoeus`. For example:
 
 ``` ruby
-HTML::Proofer.new("out/", {:ext => ".htm", :typhoeus => { :verbose => true, :ssl_verifyhost => 2 } })
+HTMLProofer.new("out/", {:ext => ".htm", :typhoeus => { :verbose => true, :ssl_verifyhost => 2 } })
 ```
 
-This sets `HTML::Proofer`'s extensions to use _.htm_, gives Typhoeus a configuration for it to be verbose, and use specific SSL settings. Check [the Typhoeus documentation](https://github.com/typhoeus/typhoeus#other-curl-options) for more information on what options it can receive.
+This sets `HTMLProofer`'s extensions to use _.htm_, gives Typhoeus a configuration for it to be verbose, and use specific SSL settings. Check [the Typhoeus documentation](https://github.com/typhoeus/typhoeus#other-curl-options) for more information on what options it can receive.
 
 You can similarly pass in a `:hydra` option with a hash configuration for Hydra.
 
@@ -218,7 +218,7 @@ The default value is `{ :typhoeus => { :followlocation => true }, :hydra => { :m
 [Parallel](https://github.com/grosser/parallel) is used to speed internal file checks. You can pass in any of its options with the options namespace `:parallel`. For example:
 
 ``` ruby
-HTML::Proofer.new("out/", {:ext => ".htm", :parallel => { :in_processes => 3} })
+HTMLProofer.new("out/", {:ext => ".htm", :parallel => { :in_processes => 3} })
 ```
 
 In this example, `:in_processes => 3` is passed into Parallel as a configuration option.
@@ -227,7 +227,7 @@ In this example, `:in_processes => 3` is passed into Parallel as a configuration
 
 Checking external URLs can slow your tests down. If you'd like to speed that up, you can enable caching for your external links. Caching simply means to skip links that are valid for a certain period of time.
 
-While running tests, HTML::Proofer will always write to a log file within a directory called *tmp/.htmlproofer*. You should probably ignore this folder in your version control system. You can enable caching for this log file by passing in the option `:cache`, with a hash containing a single key, `:timeframe`. `:timeframe` defines the length of time the cache will be used before the link is checked again. The format of `:timeframe` is a number followed by a letter indicating the length of time. For example:
+While running tests, HTMLProofer will always write to a log file within a directory called *tmp/.htmlproofer*. You should probably ignore this folder in your version control system. You can enable caching for this log file by passing in the option `:cache`, with a hash containing a single key, `:timeframe`. `:timeframe` defines the length of time the cache will be used before the link is checked again. The format of `:timeframe` is a number followed by a letter indicating the length of time. For example:
 
 * `M` means months
 * `w` means weeks
@@ -261,7 +261,7 @@ HTML-Proofer can be as noisy or as quiet as you'd like. There are two ways to lo
 
 ## Custom tests
 
-Want to write your own test? Sure! Just create two classes--one that inherits from `HTML::Proofer::Checkable`, and another that inherits from `HTML::Proofer::CheckRunner`.
+Want to write your own test? Sure! Just create two classes--one that inherits from `HTMLProofer::Checkable`, and another that inherits from `HTMLProofer::CheckRunner`.
 
 The `CheckRunner` subclass must define one method called `run`. This is called on your content, and is responsible for performing the validation on whatever elements you like. When you catch a broken issue, call `add_issue(message)` to explain the error.
 
@@ -270,7 +270,7 @@ The `Checkable` subclass defines various helper methods you can use as part of y
 Here's an example custom test that protects against `mailto` links that point to `octocat@github.com`:
 
 ``` ruby
-class OctocatLinkCheck < ::HTML::Proofer::Checkable
+class OctocatLinkCheck < ::HTMLProofer::Checkable
   def mailto?
     return false if @data_ignore_proofer || @href.nil? || @href.empty?
     return @href.match /^mailto\:/
@@ -282,7 +282,7 @@ class OctocatLinkCheck < ::HTML::Proofer::Checkable
 
 end
 
-class MailToOctocat < ::HTML::Proofer::CheckRunner
+class MailToOctocat < ::HTMLProofer::CheckRunner
   def run
     @html.css('a').each do |node|
       link = OctocatLinkCheck.new(node, self)
@@ -303,7 +303,7 @@ end
 To ignore certificates, turn off Typhoeus' SSL verification:
 
 ``` ruby
-HTML::Proofer.new("out/", {
+HTMLProofer.new("out/", {
   :typhoeus => {
     :ssl_verifypeer => false,
     :ssl_verifyhost => 0}
@@ -315,7 +315,7 @@ HTML::Proofer.new("out/", {
 To change the User-Agent used by Typhoeus:
 
 ``` ruby
-HTML::Proofer.new("out/", {
+HTMLProofer.new("out/", {
   :typhoeus => {
     :headers => { "User-Agent" => "Mozilla/5.0 (compatible; My New User-Agent)" }
 }}).run
@@ -326,7 +326,7 @@ HTML::Proofer.new("out/", {
 To exclude urls using regular expressions, include them between forward slashes and don't quote them:
 
 ``` ruby
-HTML::Proofer.new("out/", {
+HTMLProofer.new("out/", {
   :url_ignore => [/example.com/],
 }).run
 ```
