+++ bin/htmlproof
@@ -7,7 +7,7 @@ require 'html/proofer'
 require 'mercenary'
 
 Mercenary.program(:htmlproof) do |p|
-  p.version HTML::Proofer::VERSION
+  p.version HTMLProofer::VERSION
   p.description %(Test your rendered HTML files to make sure they're accurate.)
   p.syntax 'htmlproof PATH [options]'
 
@@ -44,7 +44,7 @@ Mercenary.program(:htmlproof) do |p|
     # prepare everything to go to proofer
     p.options.select { |o| !opts[o.config_key].nil? }.each do |option|
       if option.return_type.to_s == 'Array' # TODO: is_a? doesn't work here?
-        opts[option.config_key] = opts[option.config_key].map { |i| HTML::Proofer::Configuration.to_regex?(i) }
+        opts[option.config_key] = opts[option.config_key].map { |i| HTMLProofer::Configuration.to_regex?(i) }
       end
       options[option.config_key.to_sym] = opts[option.config_key]
     end
@@ -72,6 +72,6 @@ Mercenary.program(:htmlproof) do |p|
 
     path = path.delete(' ').split(',') if opts['as_links']
 
-    HTML::Proofer.new(path, options).run
+    HTMLProofer.new(path, options).run
   end
 end
