+++ lib/html-proofer/check_runner/issue.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-class HTML::Proofer::CheckRunner
+class HTMLProofer::CheckRunner
 
   class Issue
     attr_reader :path, :desc, :status, :line_number
