+++ lib/html-proofer/utils.rb
