+++ lib/html-proofer/version.rb
