+++ lib/html-proofer/xpathfunctions.rb
