+++ spec/html/proofer/command_spec.rb
@@ -93,7 +93,7 @@ describe 'Command test' do
   end
 
   it 'has every option' do
-    config_keys = HTML::Proofer::Configuration::PROOFER_DEFAULTS.keys
+    config_keys = HTMLProofer::Configuration::PROOFER_DEFAULTS.keys
     bin_file = File.read('bin/htmlproof')
     help_output = make_bin('--help')
     readme = File.read('README.md')
